% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{a data set of fish}
\format{
A data frame with 144 rows and 6 variables:
\describe{
  \item{DDT}{DDT of fish}
  \item{LENGTH}{LENGTH of fish}
  \item{MILE}{MILE of fish}
  \item{RIVER}{name of RIVER}
  \item{SPECIES}{SPECIES of fish}
  \item{WEIGHT}{WEIGHT of fish}
  ...
}
}
\source{
textbook
}
\usage{
ddt
}
\description{
144 line of measurements on fish
}
\keyword{datasets}
