% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Don't know what it is}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{physical length}
  \item{DAMAGE}{loss or harm resulting from injury to person, property, or reputation}
  ...
}
}
\usage{
fire
}
\description{
no idea
}
\keyword{datasets}
