% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot2}
\alias{myboot2}
\title{myboot}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5)
}
\arguments{
\item{iter}{number}

\item{x}{vector}

\item{fun}{function}

\item{alpha}{number}

\item{cx}{numer}
}
\description{
myboot
}
\examples{
 \dontrun{myboot2(10000,x=sam,fun="mean",alpha=0.05,xlab="mean",col="Purple",cx=1.5)}
}
