% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Title mycltp}
\usage{
mycltp(n, iter, lambda = 10)
}
\arguments{
\item{n}{number}

\item{iter}{iters}

\item{lambda}{lambda1}
}
\description{
Title mycltp
}
\examples{
 \dontrun{mycltp(1,1,1)}
}
