% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{filtering data}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{data frame}

\item{SPECIES}{condition for filter}
}
\value{
filtered data frame
}
\description{
filtering data
}
\examples{
\dontrun{myddt(ddt,SPECIES = "CCATFISH")}
}
