% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{my max lik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function}

\item{x}{vector}

\item{param}{param}

\item{...}{OTHER INPUT}
}
\description{
my max lik
}
\examples{
 \dontrun{mymaxlik(x=c(9),param=seq(0,1,length=1000),lfun=logbin)}
}
