% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymult.R
\name{mymult}
\alias{mymult}
\title{multinomial}
\usage{
mymult(iter = 100, n = 10, p = c(1, 1, 1, 1)/4)
}
\arguments{
\item{iter}{vector of numbers}

\item{n}{vector of numbers}

\item{p}{vector of numbers}
}
\value{

}
\description{
multinomial
}
\examples{
mymult(iter=1000,n=10,p=c(1,2,3,4,2)/12)
}
