% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{my curve}
\usage{
myncurve(x, mu, sigma)
}
\arguments{
\item{x}{a vector of numbers}

\item{mu}{mean}

\item{sigma}{sd}
}
\description{
my curve
}
\examples{
myncurve(5,5,5)
}
