% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{scatter histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{a vector of numbers}

\item{y}{a vector of numbers}

\item{xlab}{X label}

\item{ylab}{Y label}
}
\value{
a plot
}
\description{
scatter histogram
}
\examples{
 \dontrun{with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))}

}
