% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{DDT}{DDT of fish}
  \item{LENGTH}{LENGTH of fish}
  \item{MILE}{MILE of fish}
  \item{RIVER}{name of RIVER}
  \item{SPECIES}{SPECIES of fish}
  \item{WEIGHT}{WEIGHT of fish}
  ...
}
}
\source{
textbook
}
\usage{
ddt
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
