% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymult.R
\name{mymult}
\alias{mymult}
\title{mymult}
\usage{
mymult(iter = 100, n = 10, p = c(1, 1, 1, 1)/4)
}
\arguments{
\item{iter}{a vector of numbers}

\item{n}{a vector of numbers}

\item{p}{a vector of numbers}
}
\description{
mymult
}
\examples{
mymult(iter=1000,n=10,p=c(1,2,3,4,2)/12)
}
